Split out the login/logout/signup commands again
We'll keep the categorization though :)
