+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Add `projects create` command to create new projects
 - Add `projects update` command to update an existing project
+- Group commands into categories for better help output
 
 ### Fixed
